Fix: Replace 'volatile int' with atomic_int in debug log check
Fixes issue #1303 (compiler ewarning)

The debug logging macros typically limit the number of times a debug
message can be output, to avoid permanently overwhelming the system
with debug logging. The scoped counter used is now an atomic_int
rather than a 'volatile int'.

Note that the use of the volatile keyword in this fashion is rightly
deprecated in C++20.